Make DueDateControl
